Implement manual synchronization tests for Display <-> ImagePlus
The test and corresponding functions in LegacyImageMap test one way
synchronizations between Displays and ImagePlus.

There is a commented out test to do the same for Dataset and ImagePlus,
but I have yet to find a way to extract either the corresponding Display
or ImagePlus from a dataset, and so at present it doesn't work.